Introduct Actions::add_filter_during()
Add a filter to run between a specified start action and stop
at a specified end action.

Used to filter only during a specific stack or section of code.
Useful for targeting things like a widget or a template but may
be use when any starting and endpoint point is avialble via actions.