Support the 'block_editor' argument in extended-cpt version 4.4.0
Force the use of the block editor for this post type.
Must be used in combination with the `show_in_rest` argument.
The primary use of this argument is to prevent the block editor from
being used by setting it to false when `show_in_rest` is set to true.