Introduce `get_extended_properties` method on `Mutator_Trait
Support accessing an objects extended properties by providing
a method which returns a list of available ones.
Provides a simplified way to access properties on WP objects
which are only accessiable via the objects __get method and
don't exist directly on the object.

Implement this new logic on the `Site_Trait` and fix the phpdocs
@methods on that trait.