Fix selectively registering REST fields
Previously the first field would determine if we can use selective
registering because it would generate the box. Then if the box was
set to true by any field, the next field would not be selective.

This is now fixed by storing the original `show_in_rest` value then
looping through all fields before adding any to the box.