Return early from `use_cdn_for_resources` if incorrect action used
`wp_degister_script` will throw an error if we are in the admin
and not within the `admin_enqueue_scripts` action. Due to some
of the replaced scripts being required for the admin to load.

We simply return early and do not replace the scripts if in_admin()
and not in the correct action.