Provide post id on CMB2 field class during update_meta_value
When using a santize callback it's often useful to know the post
being updated. As the `sanitize_meta_value` WP core function does
not pass the post id, we use an existing property on the Field
class to set and pass it along.

Solves the inconsitency when the post id is available when updating
via the WP admin, but not avaialble when updating elsewhere.