Cover more cases for Meta repo's `delete_cb`
When a field is of type that supports meta, just the "delete_{type}_meta"
action so all cases of the field being deleted are sent to the callback.

Then a field is not a meta type, use the cmb2_override_{$id}_meta_remove
action to fire when a meta box is updated.

For non-meta fields we handle the callback within the repo.