Introduce Resources::get_file_modified_time method
Get the modified time of a script/style based on its url.

Used to pass a version to browsers to tell them if a file has
changed based on last time a file was modified.

Keeps a file in browser cache longer if a file revision changes
frequently, but a file is not modified often by the deployment
process nor content changes.

1. Keeps file in browser cache forever until it changes.
2. 16x faster than `Resources::get_content_hash`.
3. Pointless if `shortCssClasses` are enabled because resource files
   change on every release.

Preferred method if not using `shortCssClasses` and releases occur
much more frequently than this file is modified by release or
other processes.