Allow block CSS to load enqueue_block_assets fires on front-end
In the editor, the block.css will always load before the theme
styles so to have an accurate picture of blocks in the post editor
we must also load in the same order on the front-end.

Adds some complexity to block specificity because loading blocks.css
first means overriding selectors with same specificity ranking, but
provides the most accurate picture possible while editing.