Prevent WP core from breaking sites by using non-existent versions of resources
We say in WP 6.7.2 that WP does not follow actual library versioning when they
see fit.

We can't predict what WP will do next, so we gracefully fail when the script
cannot be found in unpkg.

Cache the failure for 1 day, so we don't continually request the meta for the
resource.