Disable WP core template output buffering for scripts
We are already using the filters to load block assets on demand.
The only benefit to the output buffering is moving styles to head instead
of the footer which prevents FOUC but slows down TTFB and adds more
blocking resources.

Using priority 5 for easy overrides should a site need to.

https://make.wordpress.org/core/2025/11/18/wordpress-6-9-frontend-performance-field-guide/#load-block-styles-on-demand-in-classic-themes